requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
Flask-Session==0.5.4
google-auth-oauthlib==0.4.5
google-auth-httplib2==0.1.1
google-api-python-client==2.12.0
dateparser==1.1.8
